split god-file agent.rs to smaller files
